+++ amethyst_ui/src/button/builder.rs
@@ -1,4 +1,5 @@
 use shred::SystemData;
+use smallvec::SmallVec;
 
 use amethyst_assets::{AssetStorage, Loader};
 use amethyst_audio::SourceHandle;
@@ -9,8 +10,9 @@ use amethyst_core::{
 use amethyst_renderer::{Texture, TextureHandle};
 
 use crate::{
-    font::default::get_default_font, Anchor, FontAsset, FontHandle, Interactable, OnUiActionImage,
-    OnUiActionSound, Selectable, Stretch, UiButton, UiText, UiTransform,
+    font::default::get_default_font, Anchor, FontAsset, FontHandle, Interactable, Stretch, Selectable,
+    UiButton, UiButtonAction, UiButtonActionRetrigger, UiButtonActionType, UiButtonActionType::*,
+    UiPlaySoundAction, UiSoundRetrigger, UiText, UiTransform,
 };
 
 use std::marker::PhantomData;
@@ -35,8 +37,8 @@ pub struct UiButtonBuilderResources<'a, G: PartialEq + Send + Sync + 'static> {
     text: WriteStorage<'a, UiText>,
     transform: WriteStorage<'a, UiTransform>,
     button: WriteStorage<'a, UiButton>,
-    action_image: WriteStorage<'a, OnUiActionImage>,
-    action_sound: WriteStorage<'a, OnUiActionSound>,
+    sound_retrigger: WriteStorage<'a, UiSoundRetrigger>,
+    button_action_retrigger: WriteStorage<'a, UiButtonActionRetrigger>,
     selectables: WriteStorage<'a, Selectable<G>>,
 }
 
@@ -58,13 +60,15 @@ pub struct UiButtonBuilder<G> {
     font_size: f32,
     image: Option<TextureHandle>,
     parent: Option<Entity>,
-    hover_image: Option<TextureHandle>,
-    hover_text_color: Option<[f32; 4]>,
-    press_image: Option<TextureHandle>,
-    press_text_color: Option<[f32; 4]>,
-    hover_sound: Option<SourceHandle>,
-    press_sound: Option<SourceHandle>,
-    release_sound: Option<SourceHandle>,
+    on_click_start_sound: Option<UiPlaySoundAction>,
+    on_click_stop_sound: Option<UiPlaySoundAction>,
+    on_hover_sound: Option<UiPlaySoundAction>,
+    // SetTextColor and SetTexture can occur on click/hover start,
+    // Unset for both on click/hover stop, so we only need 2 max.
+    on_click_start: SmallVec<[UiButtonActionType; 2]>,
+    on_click_stop: SmallVec<[UiButtonActionType; 2]>,
+    on_hover_start: SmallVec<[UiButtonActionType; 2]>,
+    on_hover_stop: SmallVec<[UiButtonActionType; 2]>,
     _phantom: PhantomData<G>,
 }
 
@@ -86,13 +90,13 @@ impl<G> Default for UiButtonBuilder<G> {
             font_size: 32.,
             image: None,
             parent: None,
-            hover_image: None,
-            hover_text_color: None,
-            press_image: None,
-            press_text_color: None,
-            hover_sound: None,
-            press_sound: None,
-            release_sound: None,
+            on_click_start_sound: None,
+            on_click_stop_sound: None,
+            on_hover_sound: None,
+            on_click_start: smallvec![],
+            on_click_stop: smallvec![],
+            on_hover_start: smallvec![],
+            on_hover_stop: smallvec![],
             _phantom: PhantomData,
         }
     }
@@ -195,43 +199,47 @@ impl<G: PartialEq + Send + Sync + 'static> UiButtonBuilder<G> {
 
     /// Text color to use when the mouse is hovering over this button
     pub fn with_hover_text_color(mut self, text_color: [f32; 4]) -> Self {
-        self.hover_text_color = Some(text_color);
+        self.on_hover_start.push(SetTextColor(text_color));
+        self.on_hover_stop.push(UnsetTextColor(text_color));
         self
     }
 
     /// Set text color when the button is pressed
     pub fn with_press_text_color(mut self, text_color: [f32; 4]) -> Self {
-        self.press_text_color = Some(text_color);
+        self.on_click_start.push(SetTextColor(text_color));
+        self.on_click_stop.push(UnsetTextColor(text_color));
         self
     }
 
     /// Button image to use when the mouse is hovering over this button
     pub fn with_hover_image(mut self, image: TextureHandle) -> Self {
-        self.hover_image = Some(image);
+        self.on_hover_start.push(SetTexture(image.clone()));
+        self.on_hover_stop.push(UnsetTexture(image));
         self
     }
 
     /// Button image to use when this button is pressed
     pub fn with_press_image(mut self, image: TextureHandle) -> Self {
-        self.press_image = Some(image);
+        self.on_click_start.push(SetTexture(image.clone()));
+        self.on_click_stop.push(UnsetTexture(image));
         self
     }
 
     /// Sound emitted when this button is hovered over
     pub fn with_hover_sound(mut self, sound: SourceHandle) -> Self {
-        self.hover_sound = Some(sound);
+        self.on_hover_sound = Some(UiPlaySoundAction(sound));
         self
     }
 
     /// Sound emitted when this button is pressed
     pub fn with_press_sound(mut self, sound: SourceHandle) -> Self {
-        self.press_sound = Some(sound);
+        self.on_click_start_sound = Some(UiPlaySoundAction(sound));
         self
     }
 
     /// Sound emitted when this button is released
     pub fn with_release_sound(mut self, sound: SourceHandle) -> Self {
-        self.release_sound = Some(sound);
+        self.on_click_stop_sound = Some(UiPlaySoundAction(sound));
         self
     }
 
@@ -239,6 +247,52 @@ impl<G: PartialEq + Send + Sync + 'static> UiButtonBuilder<G> {
     pub fn build(mut self, mut res: UiButtonBuilderResources<'_, G>) -> Entity {
         let mut id = self.name.clone();
         let image_entity = res.entities.create();
+
+        if !self.on_click_start.is_empty()
+            || !self.on_click_stop.is_empty()
+            || !self.on_hover_start.is_empty()
+            || !self.on_hover_stop.is_empty()
+        {
+            let retrigger = UiButtonActionRetrigger {
+                on_click_start: actions_with_target(
+                    &mut self.on_click_start.into_iter(),
+                    &image_entity,
+                ),
+                on_click_stop: actions_with_target(
+                    &mut self.on_click_stop.into_iter(),
+                    &image_entity,
+                ),
+                on_hover_start: actions_with_target(
+                    &mut self.on_hover_start.into_iter(),
+                    &image_entity,
+                ),
+                on_hover_stop: actions_with_target(
+                    &mut self.on_hover_stop.into_iter(),
+                    &image_entity,
+                ),
+            };
+
+            res.button_action_retrigger
+                .insert(image_entity, retrigger)
+                .expect("Unreachable: Inserting newly created entity");
+        }
+
+        if self.on_click_start_sound.is_some()
+            || self.on_click_stop_sound.is_some()
+            || self.on_hover_sound.is_some()
+        {
+            let retrigger = UiSoundRetrigger {
+                on_click_start: self.on_click_start_sound,
+                on_click_stop: self.on_click_stop_sound,
+                on_hover_start: self.on_hover_sound,
+                on_hover_stop: None,
+            };
+
+            res.sound_retrigger
+                .insert(image_entity, retrigger)
+                .expect("Unreachable: Inserting newly created entity");
+        }
+
         res.transform
             .insert(
                 image_entity,
@@ -309,30 +363,11 @@ impl<G: PartialEq + Send + Sync + 'static> UiButtonBuilder<G> {
             .insert(
                 image_entity,
                 UiButton {
-                    normal_text_color: self.text_color,
-                    hover_text_color: self.hover_text_color,
-                    press_text_color: self.press_text_color,
+                    text_color: self.text_color,
                 },
             )
             .expect("Unreachable: Inserting newly created entity");
-        if self.hover_image.is_some() || self.press_image.is_some() {
-            res.action_image
-                .insert(
-                    image_entity,
-                    OnUiActionImage::new(Some(image_handle), self.hover_image, self.press_image),
-                )
-                .expect("Unreachable: Inserting newly created entity");
-        }
 
-        if self.hover_sound.is_some() || self.press_sound.is_some() || self.release_sound.is_some()
-        {
-            res.action_sound
-                .insert(
-                    image_entity,
-                    OnUiActionSound::new(self.hover_sound, self.press_sound, self.release_sound),
-                )
-                .expect("Unreachable: Inserting newly created entity");
-        }
         image_entity
     }
 
@@ -340,4 +375,16 @@ impl<G: PartialEq + Send + Sync + 'static> UiButtonBuilder<G> {
     pub fn build_from_world(self, world: &World) -> Entity {
         self.build(UiButtonBuilderResources::<G>::fetch(&world.res))
     }
+}
+
+fn actions_with_target<I>(actions: I, target: &Entity) -> Vec<UiButtonAction>
+where
+    I: Iterator<Item = UiButtonActionType>,
+{
+    actions
+        .map(|action| UiButtonAction {
+            target: target.clone(),
+            event_type: action,
+        })
+        .collect()
 }