+++ docs/CHANGELOG.md
@@ -13,10 +13,12 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 * Introduce `application_dir` utility ([#1213])
 * Derive `Copy`, `PartialEq`, `Eq`, `Serialize`, `Deserialize` for `Flipped` component. ([#1237])
+* A way to change the default `Source` using `set_default_source` and `with_default_source`. ([#1256])
 
 ### Changed
 
 * Make `application_root_dir` return a `Result<Path>` instead of a `String` ([#1213])
+* Remove unnecessary texture coordinates offset in `Sprite::from_pixel_values` ([#1267])
 
 ### Removed
 
@@ -24,6 +26,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 [#1213]: https://github.com/amethyst/amethyst/pull/1213
 [#1237]: https://github.com/amethyst/amethyst/pull/1237
+[#1256]: https://github.com/amethyst/amethyst/pull/1256
+[#1267]: https://github.com/amethyst/amethyst/pull/1267
 
 ## [0.10.0] - 2018-12
 
@@ -46,6 +50,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Added capabilities for the `DrawFlat2D` pass to draw `TextureHandle`s by themselves. Also added a simple example for this. ([#1153])
 * Added a `Flipped` component which allows flipping sprites or images horizontally and vertically. ([#1153])
 * Added transform constructor function `Transform::new()`. ([#1187])
+* Implement generic `EventRetriggerSystem`, which enables dispatching new events as a reaction to other events ([#1189])
 
 ### Changed
 
@@ -63,6 +68,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * `BasicScenePrefab` deserialization now returns an error on invalid fields. ([#1164])
 * Reordered arguments for `Transform::set_rotation_euler` to match nalgebra's Euler angles. ([#1052])
 * Remove lifetimes from `SimpleState` ([#1198])
+* Button interactions are now handled through an `EventRetriggerSystem`, specifically hover/click sounds and image/color changes ([#1189])
 
 ### Removed
 
@@ -71,6 +77,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * `amethyst::core::Orientation` has been removed because of limited use. ([#1066])
 * `TimedDestroySystem` has been split into `DestroyAtTimeSystem` and `DestroyInTimeSystem`. ([#1129])
 * Reverted [MacOS OpenGL workaround][#972] in favor of the upstream fix in `glutin`. ([#1184])
+* `OnUiActionImage` and `OnUiActionSound` have been removed as they now work through `EventRetrigger`s ([#1189])
 
 ### Fixed
 
@@ -107,6 +114,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#1188]: https://github.com/amethyst/amethyst/pull/1188
 [#1198]: https://github.com/amethyst/amethyst/pull/1198
 [#1224]: https://github.com/amethyst/amethyst/pull/1224
+[#1189]: https://github.com/amethyst/amethyst/pull/1189
 [winit_018]: https://github.com/tomaka/winit/blob/v0.18.0/CHANGELOG.md#version-0180-2018-11-07
 [glutin_019]: https://github.com/tomaka/glutin/blob/master/CHANGELOG.md#version-0190-2018-11-09